refactor: wrap list_items in server
