+++ src/libs/notion.ts
@@ -1,3 +1,5 @@
+import { randomUUID } from "crypto";
+
 import {
   Client,
   isNotionClientError,
@@ -12,6 +14,7 @@ import type {
   BlockObjectComponent,
   BookmarkBlockObjectComponent,
   BreadcrumbBlockObjectComponent,
+  BulletedListBlockObjectComponent,
   BulletedListItemBlockObjectComponent,
   CalloutBlockObjectComponent,
   ChildDatabaseBlockObjectComponent,
@@ -30,6 +33,7 @@ import type {
   LinkPreviewBlockObjectComponent,
   LinkToPageBlockObjectComponent,
   ListBlockChildrenResponseResults,
+  NumberedListBlockObjectComponent,
   NumberedListItemBlockObjectComponent,
   ParagraphBlockObjectComponent,
   PdfBlockObjectComponent,
@@ -152,7 +156,87 @@ export const resolveBlockChildren = async (
     )
   );
   const nonNullBlockObjectComponents = blockObjectComponents.filter(notNull);
-  return nonNullBlockObjectComponents;
+  return wrapListItems(nonNullBlockObjectComponents);
+};
+
+export const wrapListItems = (
+  blocks: Array<BlockObjectComponent>
+): Array<BlockObjectComponent> => {
+  return blocks.reduce(
+    (
+      prevList: Array<BlockObjectComponent>,
+      currBlock: BlockObjectComponent
+    ): Array<BlockObjectComponent> => {
+      /* If the block.type is neither
+       * 'bulleted_list_item' nor 'numbered_list_item' nor 'bulleted_list' nor 'numbered_list',
+       * return the block as it is.
+       */
+      if (
+        currBlock.type !== "bulleted_list" &&
+        currBlock.type !== "numbered_list" &&
+        currBlock.type !== "bulleted_list_item" &&
+        currBlock.type !== "numbered_list_item"
+      ) {
+        return [...prevList, currBlock];
+      }
+      const prevBlock = prevList[prevList.length - 1];
+
+      // For the first time,
+      // create a list object and add list_item to the list.
+      const id = randomUUID();
+      if (
+        // bulleted_list_item
+        currBlock.type === "bulleted_list_item" &&
+        prevBlock?.type !== "bulleted_list"
+      ) {
+        const bulletedList = {
+          id,
+          type: "bulleted_list",
+          bulleted_list: { items: [currBlock] },
+        } satisfies BulletedListBlockObjectComponent;
+        return [...prevList, bulletedList];
+      }
+      if (
+        // numbered_list_item
+        currBlock.type === "numbered_list_item" &&
+        prevBlock?.type !== "numbered_list"
+      ) {
+        const numberedList = {
+          id,
+          type: "numbered_list",
+          numbered_list: { items: [currBlock] },
+        } satisfies NumberedListBlockObjectComponent;
+        return [...prevList, numberedList];
+      }
+
+      // Intermediate,
+      // add the second and subsequent list items to the previous list.
+      if (
+        // bulleted_list
+        currBlock.type === "bulleted_list_item" &&
+        prevBlock?.type === "bulleted_list"
+      ) {
+        prevBlock.bulleted_list.items = [
+          ...prevBlock.bulleted_list.items,
+          currBlock,
+        ];
+      }
+      if (
+        // numbered_list_item
+        currBlock.type === "numbered_list_item" &&
+        prevBlock?.type === "numbered_list"
+      ) {
+        prevBlock.numbered_list.items = [
+          ...prevBlock.numbered_list.items,
+          currBlock,
+        ];
+      }
+
+      // If the condition is not met, do not display/render.
+      return prevList;
+    },
+    []
+  );
 };
 
 export const fetchBlockComponents = async (blockId: string) => {
